Use floating point for physics and collision systems

+++ Src/ECScape.Core/Components/Collectable.cs
@@ -3,5 +3,5 @@
 internal sealed class Collectable : IComponent
 {
     public TimeSpan CollectInterval { get; init; } = TimeSpan.FromSeconds(1);
-    public int ScoreAmount { get; init; } = 1;
+    public double ScoreAmount { get; init; } = 1;
 }
